敵のひるみ処理を追加

+++ Source/Mame/Game/EnemyGolem.cpp
@@ -50,7 +50,7 @@ EnemyGolem::EnemyGolem()
         GetStateMachine()->SetState(static_cast<UINT>(StateMachineState::IdleState));
     }
 
-    SetType(static_cast<UINT>(Enemy::TYPE::Golem));
+    SetType(Enemy::TYPE::Golem);
 
     // ImGui–¼‘Oİ’è
     SetName("EnemyGolem" + std::to_string(nameNum_++));
