+++ Source/Mame/Game/EnemyGolemState.cpp
@@ -57,7 +57,7 @@ namespace EnemyGolemState
         owner->PlayAnimation(static_cast<UINT>(EnemyGolem::Animation::Landing), false);
 
         landingTimer = 0.0f;
-        
+
         timer = 0.0f;
         isCameraShake = false;
     }
@@ -299,7 +299,7 @@ namespace EnemyGolemState
     // 初期化
     void Attack1State::Initialize()
     {
-        
+
 
         owner->SetCurrentState(static_cast<UINT>(EnemyGolem::StateMachineState::Attack1State));
 
@@ -318,7 +318,7 @@ namespace EnemyGolemState
 
     // 更新
     void Attack1State::Update(const float& elapsedTime)
-    {        
+    {
         // 腕ひき
         if (!isAttack1_tame)
         {
@@ -407,7 +407,7 @@ namespace EnemyGolemState
         isComboAttack3Up        = false;
         isComboAttack3Down      = false;
         isComboAttack3Return    = false;
-        
+
         moveTimer = 0.0f;
         moveFrontTimer = 0.0f;
 
@@ -427,7 +427,7 @@ namespace EnemyGolemState
 
         // 待機ステートへ
         if (owner->comboAttackStone->isChangeState)
-        {   
+        {
             owner->GetStateMachine()->ChangeState(static_cast<UINT>(EnemyGolem::StateMachineState::IdleState));
             owner->comboAttackStone->isChangeState = false;
         }
@@ -583,7 +583,7 @@ namespace EnemyGolemState
         {
             // アニメーションが終わったら
             if (!owner->IsPlayAnimation())
-            {                
+            {
             }
         }
     }
@@ -605,7 +605,7 @@ namespace EnemyGolemState
 
         return false;
     }
-    
+
     // 回転処理
     void ComboAttack1State::Turn(const float& elapsedTime)
     {
@@ -646,7 +646,7 @@ namespace EnemyGolemState
                 isDown = true;
             }
         }
-        
+
         if (!isReturn && isDown)
         {
             getUpTimer += elapsedTime;
@@ -709,7 +709,7 @@ namespace EnemyGolemState
             AttackUpdate(elapsedTime);
         }
         else
-        {   
+        {
             if (delayTimer >= maxDelayTime)
             {
                 // 待機ステート
@@ -813,7 +813,7 @@ namespace EnemyGolemState
             if (setState == static_cast<UINT>(STATE::Summon))
             {
                 std::vector<Enemy*> enemy;
-                EnemyManager::Instance().GetSpecifyEnemy(static_cast<UINT>(Enemy::TYPE::Normal), enemy);
+                EnemyManager::Instance().GetSpecifyEnemy(Enemy::TYPE::Normal, enemy);
 
                 if (enemy.size() > 0)
                 {
@@ -846,7 +846,7 @@ namespace EnemyGolemState
             owner->GetStateMachine()->ChangeState(static_cast<UINT>(EnemyGolem::StateMachineState::RoarState));
             break;
         }
-        
+
         isState[setState] = true;
     }
 
@@ -936,7 +936,7 @@ namespace EnemyGolemState
             true);
 
         owner->model->weight = 0.0f;
-        
+
         // 変数初期化
         isChangeState = false;
         moveSpeed = 0.0f;
@@ -1003,7 +1003,7 @@ namespace EnemyGolemState
         vec = Normalize(vec);
         vec = vec * moveSpeed * elapsedTime;
         vec.y = 0.0f;   // Yは移動なし。
-        
+
         owner->GetTransform()->AddPosition(vec);
     }
 
@@ -1031,7 +1031,7 @@ namespace EnemyGolemState
 
 
         // 回転処理
-        Turn(elapsedTime);  
+        Turn(elapsedTime);
     }
 
     // 終了化
